feat: rtk 로직 수정
