+++ src/pages/main/components/ItemList.js
@@ -1,7 +1,6 @@
 import OneItem from './OneItem'
 import styled from 'styled-components'
 
-
 const ItemList = ({ data }) => {
 	console.log(data)
