+++ src/pages/main/index.js
@@ -1,26 +1,15 @@
 import { useEffect, useState } from 'react'
 import ItemList from './components/ItemList'
-import { mainApi } from '../../apis/MainApi'
-
+import { useDispatch } from 'react-redux'
+import { getIssue } from '../../reducer/issue'
 const MainPage = () => {
-	const [data, setData] = useState()
+	const dispatch = useDispatch()
 	const [isLoading, setIsLoading] = useState(true)
-
-	const useMainApi = async () => {
-		try {
-			const res = await mainApi.getApi()
-			setData(res.data)
-			console.log('getApi', res.data)
-			setIsLoading(false)
-		} catch (err) {
-			console.log(err)
-			setIsLoading(false)
-		}
-	}
 
 	useEffect(() => {
 		dispatch(getIssue())
-	}, [getIssue])
-	return !isLoading && data && <ItemList data={data}></ItemList>
+	}, [])
+
+	return !isLoading && <ItemList></ItemList>
 }
 export default MainPage
